Add method to get weak count address in core references
A new method called GetWeakCountAddress() has been added to the Ref classes in Core library. This method will return the address of the weak reference count from the atomic variable. It will primarily be used in incrementing weak count in shared references and when creating a weak reference from a shared reference.