Add OS and System specs to profiling data
Included function to identify the operating system and added system specifications to the profiling service. This will help in more advanced profiling, by providing context about the hardware and OS the program is running on. For now, the graphics card specification is hardcoded, but in the future it should be dynamically fetched.