Refactor ImGuiLayer to be abstract
The `ImGuiLayer` class has been refactored to be an abstract class, with `Begin` and `End` methods marked as pure virtual. A `Create` method has been added to create instances of the appropriate derived class based on the current graphics API. The `VulkanImGuiLayer` class has been extended with `Begin` and `End` methods and moved to the `Vulkan` directory accordingly. Also, functions related to instance's life cycle and rendering such as `OnAttach`, `OnDetach`, and `OnImGuiRender` are removed from `ImGuiLayer`.