Replace local variable with class member in VulkanSwapChain
The local variable 'imageCount' in VulkanSwapChain.cpp was replaced with a new class member 'm_imageCount'. Now, 'm_imageCount' is used to store and manipulate the number of images in the swap chain. This change is reflected in all the function calls and object references where 'imageCount' was previously used.