Implement temporary UI render fix
A temporary workaround has been introduced to enable UI rendering. Changes are made to the VulkanRenderContext, main and App files. Note that this is a temporary solution and a more optimal fix will be implemented in later updates, as current modifications might pose potential issues with the overall application flow.