Update ThryveRenderer CMakeLists and added new files
The ThryveRenderer CMakeLists file has been updated to include new files for Camera, VulkanCommandQueueManager and VulkanBuffer. The include directories and dependencies have also been condensed and tidied up for more efficient code. Furthermore, GLM definitions were added to the compile process. The project's minimum required CMake version was also lowered to 3.5, and Volk was added as a subdirectory and included in the project's directories.